decimales devolucion

+++ database/migrations/2020_06_16_000509_create_provider_return_details_table.php
@@ -26,8 +26,8 @@ class CreateProviderReturnDetailsTable extends Migration
             $table->bigInteger('meas_fk')->unsigned(); //Llave Foranea Catálogo Unidad de Medida
             $table->foreign('meas_fk')->references('meas_pk')->on('measurements')->onUpdate('cascade');
 
-            $table->integer('prrd_quantity'); //Cantidad
-            $table->integer('prrd_quantity_purchase'); //Cantidad Compra (Validar cantidad maxima)
+            $table->decimal('prrd_quantity', 12, 2); //Cantidad
+            $table->decimal('prrd_quantity_purchase', 12, 2); //Cantidad Compra (Validar cantidad maxima)
             $table->decimal('prrd_price', 12, 2); //Precio Compra
             $table->smallInteger('prrd_status')->default(1); //Estatus
             $table->timestamps();
